+++ Picture/picture-picture/src/main/java/com/lz/picture/model/dto/pictureInfo/UserPictureInfoQuery.java
@@ -50,12 +50,12 @@ public class UserPictureInfoQuery extends PageDomain implements Serializable {
     /**
      * 图片宽度
      */
-    private Long picWidth;
+    private Integer picWidth;
 
     /**
      * 图片高度
      */
-    private Long picHeight;
+    private Integer picHeight;
 
     /**
      * 宽高比例
