+++ Picture/picture-picture/src/main/java/com/lz/picture/model/vo/pictureInfo/PictureInfoTableVo.java
@@ -58,12 +58,12 @@ public class PictureInfoTableVo implements Serializable {
     /**
      * 图片宽度
      */
-    private Long picWidth;
+    private Integer picWidth;
 
     /**
      * 图片高度
      */
-    private Long picHeight;
+    private Integer picHeight;
 
     /**
      * 宽高比例
